Add Validation and Logging behaviors to MediatR
In this revision, two behavioral changes were made to the MediatR configuration in the BuberDinner module. The ValidationBehavior and LoggingBehavior were moved from being transient services and added directly in the MediatR configuration. This change was necessary to improve the application's performance and maintainability by reducing unnecessary redundancy and facilitating easier behavioral modifications in future.