Add UnitTests project and configuration files
Created a new project named UnitTests and added it to the solution. This project will serve as the dedicated space for unit testing. The Xunit framework is used for unit testing, hence added the required package references and setup. Also added related Xunit configuration files for the control of test execution. The project file includes basic settings for a .NET 6.0 project with enabled nullable reference types. Lastly, a global 'using' directive for Xunit is added to avoid repeated import in every test class.