Move IApplicationDbContext to Domain module
The IApplicationDbContext interface is repositioned from the Application module to the Domain module in order to align with the domain-centric design. The database context is more related to the domain, so it makes logical sense to keep it within the Domain module. This commit also adjusts the namespace of IApplicationDbContext in all files and moves the Microsoft.EntityFrameworkCore package reference from Application to Domain in csproj files.