Update .NET framework and improve code
Updated the .NET framework version from .NET 7.0 to .NET 8.0 in the project files and packages configuration to stay up-to-date with the latest version. Also, changed the usage of Any() to Count in the MediatorExtension.cs file for performance improvement. Modified return types in ApiController.cs to enhance code clarity. Consequently required some updates to the package versions for compatibility with the new .NET version.